Allow undef parameter passing with nil
Now, when `nil` is used as a param value, its `inspect` method is not
called to construct the Puppet parameter value. Instead, the unquoted
string `'undef'` is used in the generated resource declaration. So now
`let(:params) { { :foo => nil } }` leads to the desired resource
declaration `class { 'test': foo => undef }`.

Puppet `undef` is nearly equivalent to Ruby `nil`, and there is no `nil`
in Puppet, so it makes sense to do this translation.

Before this change, it was impossible to pass `undef` as a parameter.
Using `nil`, `:undef`, or `'undef'` all failed for various reasons.

Credit to the discussion at [1] for identifying the issue and proposing
an alternate solution.

Fixes #279.

[1] <https://groups.google.com/forum/#!topic/puppet-users/6nL2eROH8is>